Move SearchParameters to file in adapters dir
